chore: update hydra script to use env vars
